Lambda_App/requirements.txt
langchain==0.0.352
openai
sentence_transformers
faiss-cpu
line-bot-sdk